requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
lightgbm==4.0.0
xgboost==1.7.6
catboost==1.2

# Deep Learning (Optional - for LSTM models)
torch==2.0.1
transformers==4.31.0

# NLP & Sentiment Analysis
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.17.1

# FinBERT for financial sentiment
# Note: Install via: pip install git+https://github.com/ProsusAI/finBERT.git

# Data Collection APIs
praw==7.7.1  # Reddit API
tweepy==4.14.0  # Twitter API
requests==2.31.0
beautifulsoup4==4.12.2

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1
pandas-datareader==0.10.0

# Technical Analysis
ta-lib==0.4.28  # Requires system install: conda install -c conda-forge ta-lib
ta==0.11.0  # Pure Python alternative

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Backtesting
backtrader==1.9.78.123
vectorbt==0.25.4

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Data Storage
sqlalchemy==2.0.19
psycopg2-binary==2.9.7  # PostgreSQL
pymongo==4.4.1  # MongoDB (optional)

# Utilities
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1
pylint==2.17.5

# OpenAI (Optional - for GPT sentiment)
openai==0.27.8

# API Rate Limiting
ratelimit==2.2.1

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Performance
numba==0.57.1
